+++ NavitaWeb/Models/Patrimonio.cs
@@ -14,7 +14,7 @@ namespace NavitaWeb.Models
         [Required]
         public string Nome { get; set; }
         public string Descricao { get; set; }
-        public double NumeroTombo { get; set; }
+        public int NumeroTombo { get; set; }
 
         public DateTime Created { get; set; }
